Update NMF: x,y,z,r now float32, which is officially supported by HDF group

+++ btmorph2/tools/swc_to_hdf5.py
@@ -47,10 +47,10 @@ def swc_to_nmf(file_name,out_file=None):
     swc_grp = f.create_group("swc")
     swc_grp.create_dataset("index",data=ids,dtype="int32")
     swc_grp.create_dataset("type",data=s_types,dtype="int8")
-    swc_grp.create_dataset("x",data=xs,dtype="float16")
-    swc_grp.create_dataset("y",data=ys,dtype="float16")
-    swc_grp.create_dataset("z",data=zs,dtype="float16")
-    swc_grp.create_dataset("r",data=rs,dtype="float16")
+    swc_grp.create_dataset("x",data=xs,dtype="float32")
+    swc_grp.create_dataset("y",data=ys,dtype="float32")
+    swc_grp.create_dataset("z",data=zs,dtype="float32")
+    swc_grp.create_dataset("r",data=rs,dtype="float32")
     swc_grp.create_dataset("parent_index",data=parents,dtype="int32")
     # add meta-data about the SWC soma-type
     """internally used codes for soma types are btmorph specific
